Rpc: correctly marshal int, int32 and int64
I fear that something may be using it as an involuntary feature, but it is wrong. This should fix #128

+++ src/lib/rpc.ml
@@ -171,9 +171,9 @@ let rec to_string t =
 
 let rpc_of_t x = x
 
-let rpc_of_int64 i = Int i
+let rpc_of_int64 i = Int64 i
 
-let rpc_of_int32 i = Int (Int64.of_int32 i)
+let rpc_of_int32 i = Int32 (Int64.of_int32 i)
 
 let rpc_of_int i = Int (Int64.of_int i)
